Add paginated user search, contacts, suggestions API
Introduces paginated user search, recent contacts, and suggested users endpoints with new DTOs and Dapper SQL. Adds JobTitle to User (with migration). Updates repositories for paging/projection, improves DTO typing, and refactors for null safety and parallel DB calls. Adds general API rate limiting and updates app settings/constants.

+++ PutZige.Application/DTOs/Messaging/ConversationDto.cs
@@ -11,6 +11,6 @@ public sealed class ConversationDto
     public bool IsOnline { get; init; }
 
     public MessageDto? LastMessage { get; init; }
-    public int UnreadCount { get; init; }
+    public long UnreadCount { get; init; }
     public DateTime? LastActivity { get; init; }
 }
